fix: abci error code (#1280)
## Summary
This changes `AbciErrorCode` to no longer be able to represent
tendermint Code 0.

## Background
Previously `0` was defined as `UNSPECIFIED` (in line with common
practice for protobuf enum variant assignations). However, in terms of
ABCI codes, `0` should always represent success rather than an error
case.

We also had an `impl From<NonZeroU32> for AbciErrorCode` which allowed
for constructing arbitrary `AbciErrorCode`s with meaningless values in
the context of the Astria codebase. This impl was unused however.

## Changes
* Changed `AbciErrorCode` to wrap a `NonZeroU32` rather than a `u32` so
that const tendermint `Code`s matching the `AbciErrorCode` variants can
be created without using `unsafe` blocks. All such unsafe code is
restricted to the `AbciErrorCode` impl and is trivially checkable to be
safe.
* Removed the `UNSPECIFIED` variant, making the `AbciErrorCode` name
more meaningful, as it now only represents error cases.
* Removed the `From<NonZeroU32>` impl, meaning only the set of defined
const values can be used.
* Replaced `impl From<AbciErrorCode> for Code` with `pub const fn
value(self) -> NonZeroU32` so that const `Code`s can be used (useful
when matching on `Code` values).

## Testing
N/A (all new code is trivial IMO).

## Breaking Changelist
* Technically removing the pub const `AbciErrorCode::UNSPECIFIED` breaks
the astria-core crate's API, but as it's unpublished yet, and since that
const was never actually used afaik, this should not be an issue.

## Related Issues
Closes #1259.