refactor(core, sequencer)!: require that bridge unlock address always be set (#1339)
## Summary
Requires that
`astria.protocol.v1alpha1.BridgeUnlockAction.bridge_address` is always.

## Background
Before this patch, the transaction signer containing the bridge unlock
action was used as the bridge address if the bridge address was not set.
This means that setting the bridge address to the signer or leaving it
unset resultes in the same behavior. This patch makes the bridge address
a requirement so that a bridge unlock is easier to understand at a first
glance (i.e. without needing to remember implementation or spec
details), and because two separate actions should not lead to the same
result.

## Changes
- Make `astria.protocol.v1alpha1.BridgeUnlockAction.bridge_address` a
required field.

## Testing
Tests have been updated to reflect this change: explicitly setteing the
bridge address to the signer makes those tests pass again that relied on
the implicit behaviour.

## Breaking Changelist
This is a protocl breaking change because newer sequencers will reject
bridge unlock actions without a bridge address.

## Related Issues
Closes #1338