fix(ci): enable bridge withdrawer building with tag (#1374)
## Summary
Updates the image build to work on a specific tag when binary name and
image name don't match

## Background
The `astria-bridge-withdrawer` binary is pushed to an
`evm-bridge-withdrawer` docker package and the tag used is
`bridge-withdrawer-v<semantic version>` for releases. This is
potentially suboptimal but has already been done. Our image build script
doesn't work with this setup.

## Changes
- update image build to consider `binary-name` without `astria` prefix
and take a separate `package-name` for most instances these are the same
but for when they are not it supports it. There is also a
`binary-prefix` which can be set to provide more flexibility for the
future.

## Testing
CI/CD tests

## Related Issues

closes https://github.com/astriaorg/astria/issues/1373