celestia-node chart as subchart of evm-stack (#1403)
## Summary
A change to dependency level in our charts, `celestia-node` will now be
a direct sub chart of the `evm-stack`
## Background
Changes to the way our chart works broke celestia-node chart deployment
duo to dependency level issues, this PR aims to fix it.
## Changes
- moved celestia-node to be a dependency of evm-stack instead of
evm-rollup
- remove TOKEN_SERVER_URL and token server functionality
- remove token-svc

## Testing
Tested against a local k8 cluster, both with celestia-node enabled and
disabled

## Breaking changes
Non breaking but as default celestia auth is disabled, in case there is
a need for a token with read/write access user with ssh access will have
to fetch the key manually for conductor to operate properly.
## Related Issues
closes #1362