chore(sequencer): change test addresses to versions with known private keys (#1487)
## Summary
Updated public addresses in `app/test_util.rs` to addresses with known
private keys.

## Background
Needed for PR #1408. We want to separate out the 'breaking change' into
a different PR since it is unrelated.

## Breaking Changes

While the sequencer app hash snapshots are changed, this patch is not
breaking as not business logic was updated.