fix(ci): use commit SHA instead of PR number preview-env images (#1501)
## Summary
Updates preview environment manifests to use the exact commit sha when
referencing artifacts to deploy (docker images). Also adds links in the
GUI back to the PR and commit.

## Background
This is to ensure that any subsequent updates to a PR will always result
in the preview env looking for the rebuilt image (tagged with full
commit hash) rather than assuming a cached one might be valid (e.g.
tagged with PR number).This change is in addition to setting the
imagePullPolicy to Always to make the configuration explicit.

## Changes
- Update github action to tag docker images with the full commit hash

## Testing
- Manually tested by applying the preview environment manifests on dev
cluster and creating a new PR to verify images were tagged correctly and
that ArgoCD pulled the correct image.