fix: ignore `RUSTSEC-2024-0370` (#1483)
## Summary
Ignore RustSec warning.

## Background
We get a non-critical warning when running `cargo audit`:
[RUSTSEC-2024-0370](https://rustsec.org/advisories/RUSTSEC-2024-0370).

When running `cargo tree -i -p=proc-macro-error` we can see that
`proc-macro-error` is a dependency of `borsh` which is already at the
latest version.

Given that the RustSec report doesn't suggest any concrete problems with
`proc-macro-error` and how difficult it will be to move away from this
dependency, I have just ignored this warning in CI.

## Changes
- Ignore RustSec warning in `.cargo/audit.toml`.
- Also fixed a typo causing a compiler warning.

## Testing
Ran `cargo audit` locally.