feat(conductor): implement restart logic (#1463)
## Summary
Added logic to restart conductor tasks if it receives a
`PermissionDenied` code from the execution layer.

## Background
Conductor previously needed to be restarted if a rollup was restarted.
This change is meant to make it such that if a rollup goes down and
comes back online, the conductor can restart internally and
automatically as opposed to needing a manual restart.

## Changes
- Added wrapper task and corresponding handle around conductor to manage
a restart, allowing the `InnerConductorTask` to shut down gracefully
before restarting.
- Added error propagation within the conductor to ensure tasks do not
silently fail.
- Added `RestartToken` to be returned by the `InnerConductorTask` on
shutdown to indicate to the wrapper whether to restart.
- Added checks for task errors with `tonic::code::PermissionDenied` to
initiate a restart.
- Implemented logic to check a returned error for a need to restart and
function `restart()` to end the tasks and create new ones.

## Testing
Added blackbox test and helper functions to ensure conductor restarts
after receiving a `PermissionDenied` status code.

## Breaking Changelist
Not breaking, but we should have documentation that reflects the fact
that rollups need to return a `PermissionDenied` error if they receive
an `execute_block` gRPC before `get_genesis_info` or
`get_commitment_state`. Similarly, if the type of code used for this
changes, the conductor restart logic will need to be changed to reflect
this. It's quite rigid right now.

## Related Issues
closes #906