chore(ci): bump rust toolchain to 1.81 (#1523)
## Summary
Updates the rust toolchain used for all rust releated github jobs to and
the containerfile to 1.81.0.

## Background
Rust [1.81 was
released](https://blog.rust-lang.org/2024/09/05/Rust-1.81.0.html) and
comes with a load of niceties that we would like to use. Among them of
immediate interest are:

1. lint reasons (https://github.com/astriaorg/astria/issues/1521)
2. fixes `clippy::blocks_in_conditions` lint triggering on proc macros
like `tracing::instrument` (which
https://github.com/astriaorg/astria/pull/1368 needed to work around)
3. lazy lock and lazy cell (to replace once_cell) in 1.80
(https://github.com/astriaorg/astria/issues/1520)
4. checked cfg names (also 1.80 which flags a few issues in our code and
should be immediately available).

Note that this is orthgonal to setting the MSRV in individual Astria
crates.

## Changes
- Updates all github workflows to use the `1.81.0` toolchain (except for
rustfmt and dylint, which need nightlies, respectively).
- Fix all new clippy warnings.
- Upgraded to the most recent dylint to make it compile again
- Upgrade the single containerfile to use `rust:1.81-bookwarm` as the
base image

## Related Issues
Closes https://github.com/astriaorg/astria/issues/1522